refactor: move memory subsystem to codelens-core
- Create codelens-core/src/memory.rs with pure filesystem memory operations
  (list, read, write, delete, rename) independent of MCP protocol
- Simplify tools/memory.rs to thin adapter delegating to core
- Update tools/session.rs import to use codelens_core::memory
